list meals with a table on main page - create database - add MealRepository - add meals as a table in the template

+++ src/main/java/com/greenfox/balintvecsey/models/Meal.java
@@ -22,5 +22,5 @@ public class Meal {
   private Timestamp date;
   private String type;
   private String description;
-  private int calories;
+  private double calories;
 }
